Service klase

+++ src/main/java/rs/ac/uns/ftn/tseo/ssd/repository/DokumentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import rs.ac.uns.ftn.tseo.ssd.model.Dokument;
 
-public interface DokumentRepository extends JpaRepository <Dokument,Long>{
+public interface DokumentRepository extends JpaRepository <Dokument,Integer>{
 
 	Dokument findByNaziv(String naziv);
 	List <Dokument> findByTip(String tip);
