+++ src/main/java/rs/ac/uns/ftn/tseo/ssd/repository/ERacunRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import rs.ac.uns.ftn.tseo.ssd.model.ERacun;
 import rs.ac.uns.ftn.tseo.ssd.model.Uplata;
 
-public interface ERacunRepository extends JpaRepository <ERacun,Long> {
+public interface ERacunRepository extends JpaRepository <ERacun,Integer> {
 	ERacun findByUplata(Uplata uplata);
 
 }
